% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curand-package.r
\docType{package}
\name{curand-package}
\alias{curand-package}
\title{curand}
\description{
Random number generators that run on the gpu via CUDA's cuRAND
library.
}
\author{
Drew Schmidt
}
\keyword{Package}

