% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_normal.r
\name{curand_log_normal}
\alias{curand_log_normal}
\title{curand_log_normal}
\usage{
curand_log_normal(n, meanlog = 0, sdlog = 1, seed = getseed(),
  type = "double")
}
\arguments{
\item{n}{The number of values to generate}

\item{seed}{Seed for the random number generation.}

\item{type}{'double' or 'float'}

\item{mean, sd}{Parameters for normal random variables.}
}
\description{
Generate from a normal distribution using a gpu.
}

