% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.r
\name{curand_normal}
\alias{curand_normal}
\title{curand_normal}
\usage{
curand_normal(n, mean = 0, sd = 1, seed = getseed(), type = "double")
}
\arguments{
\item{n}{The number of values to generate}

\item{mean, sd}{Parameters for normal random variables.}

\item{seed}{Seed for the random number generation.}

\item{type}{'double' or 'float'}
}
\description{
Generate from a normal distribution using a gpu.
}

