% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform.r
\name{curand_uniform}
\alias{curand_uniform}
\title{curand_uniform}
\usage{
curand_uniform(n, min = 0, max = 1, seed = getseed(), type = "double")
}
\arguments{
\item{n}{The number of values to generate}

\item{min, max}{Parameters for uniform random variables.}

\item{seed}{Seed for the random number generation.}

\item{type}{'double' or 'float'}
}
\description{
Generate from a uniform distribution using a gpu.
}
\references{
CUDA Toolkit Documentation for cuRAND
\url{https://docs.nvidia.com/cuda/curand/index.html}
}

