% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexp.r
\name{rexp}
\alias{rexp}
\title{rexp}
\usage{
rexp(n, rate = 1, seed = getseed(), type = "double")
}
\arguments{
\item{n}{The number of values to generate}

\item{rate}{Parameter for exponential random variables.}

\item{seed}{Seed for the random number generation.}

\item{type}{'double' or 'float'}
}
\description{
Generate from a exponential distribution using a gpu.
}
\details{
We use the cuRAND uniform generator together with the fact that if
\eqn{X \sim U(0, 1)} then \eqn{Y = -\lambda^{-1}\log(X) \sim Exp(\lambda)}.
}
\references{
Casella, G. and Berger, R.L., 2002. Statistical inference (Vol. 2). Pacific
Grove, CA: Duxbury.

Rizzo, M.L., 2007. Statistical computing with R. Chapman and Hall/CRC.
}

