% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rweibull.r
\name{rweibull}
\alias{rweibull}
\title{rweibull}
\usage{
rweibull(n, shape, scale = 1, seed = getseed(), type = "double")
}
\arguments{
\item{n}{The number of values to generate}

\item{shape, scale}{Parameters for weibull random variables.}

\item{seed}{Seed for the random number generation.}

\item{type}{'double' or 'float'}
}
\description{
Generate from a weibull distribution using a gpu.
}
\references{
Casella, G. and Berger, R.L., 2002. Statistical inference (Vol. 2). Pacific
Grove, CA: Duxbury.

Rizzo, M.L., 2007. Statistical computing with R. Chapman and Hall/CRC.
}

