\name{curand-package}
\alias{curand-package}
\alias{curand}
\docType{package}
\title{\packageTitle{curand}}
\description{\packageDescription{curand}}
\section{Package Content}{\packageIndices{curand}}
\author{\packageAuthor{curand}}
\section{Maintainer}{\packageMaintainer{curand}}
\keyword{package}
