% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_normal.r
\name{curand_log_normal}
\alias{curand_log_normal}
\title{curand_log_normal}
\usage{
curand_log_normal(n, meanlog = 0, sdlog = 1, seed = getseed(),
  type = "double")
}
\arguments{
\item{n}{The number of values to generate}

\item{meanlog, sdlog}{Parameters for log normal random variables.}

\item{seed}{Seed for the random number generation.}

\item{type}{'double' or 'float'}
}
\description{
Generate from a log normal distribution using a gpu.
}

