% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.r
\name{curand_weibull}
\alias{curand_weibull}
\title{curand_weibull}
\usage{
curand_weibull(n, shape, scale = 1, seed = getseed(), type = "double")
}
\arguments{
\item{n}{The number of values to generate}

\item{seed}{Seed for the random number generation.}

\item{type}{'double' or 'float'}

\item{mean, sd}{Parameters for weibull random variables.}
}
\description{
Generate from a weibull distribution using a gpu.
}
\references{
Rizzo, M.L., 2007. Statistical computing with R. Chapman and
Hall/CRC.
}

